feat(planning): fractional demand from a per-member portion factor (plan 0027 P-D1/D4/D5/D6)
Member.portionFactor (default 1) and demandPortions — Σ of the eaters'
factors, the whole-number override still winning. The per-person macro
lens weights by the factor (× override ÷ Σ factors under an override) and
MealSetMacros carries its servings + demand so portionShareLine can name
the denominator ('Jun · ¾ of 1¾ portions'). CookSession portions become a
double end to end (CoveredMeal, totalPortions, the cook marker's batch),
and every portion count prints through core/units/portions.dart — ¼ ½ ¾
glyphs, never 1.75, a trimmed decimal for a non-quarter share.

The identity test runs today's fixtures — every eater/override shape —
through the new derivations with all factors at 1 and asserts the demand,
the macros under every lens, the cook plan, the nudge and the printed
strings are byte-identical (P-D6).

+++ app/lib/features/cook_plan/data/cook_plan_repository_impl.dart
@@ -75,7 +75,7 @@ class SqliteCookPlanRepository implements CookPlanRepository {
     for (final row in rows) {
       final recipeId = row['recipe_id'] as String;
       final eaters = jsonDecode(row['eaters'] as String? ?? '[]') as List;
-      final portions = (row['portions'] as int?) ?? eaters.length;
+      final portions = ((row['portions'] as int?) ?? eaters.length).toDouble();
       byRecipe[recipeId] ??= PlannedRecipe(
         recipeId: recipeId,
         title: row['title'] as String,
